Revert "flytt erEndringIUtbetaling til BeregningService"
This reverts commit f2e74e43b805030c2f4e876516f904022b2596ce.
